@@ -50,19 +50,19 @@
     ]
   },
   "dependencies": {
-    "ml-xsadd": "^1.0.0"
+    "ml-xsadd": "^2.0.0"
   },
   "devDependencies": {
-    "@types/jest": "^22.2.3",
+    "@types/jest": "^23.3.1",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.26.2",
-    "codecov": "^3.0.2",
-    "jest": "^22.4.3",
+    "codecov": "^3.0.4",
+    "jest": "^23.5.0",
     "npm-run-all": "^4.1.3",
-    "shx": "^0.2.2",
-    "ts-jest": "^22.4.6",
-    "tslint": "^5.10.0",
-    "tslint-config-prettier": "^1.12.0",
+    "shx": "^0.3.2",
+    "ts-jest": "^23.1.3",
+    "tslint": "^5.11.0",
+    "tslint-config-prettier": "^1.14.0",
     "typedoc": "^0.11.1",
-    "typescript": "^2.8.3"
+    "typescript": "^3.0.1"
   }
 }
